refactor: move group creation modal
