+++ src/actions/calendar-actions-fn.tsx
@@ -14,11 +14,11 @@ import moment from 'moment';
 import { FOLDER_OPERATIONS } from '../constants/api';
 import { getFolderRequest } from '../soap/get-folder-request';
 import { getShareInfoRequest } from '../soap/get-share-info-request';
+import { CreateGroupModal } from './modals/create-group-modal';
 import { folderAction } from '../store/actions/calendar-actions';
 import { StoreProvider } from '../store/redux';
 import { ActionsClick } from '../types/actions';
 import { NewModal } from '../view/move/new-calendar-modal';
-import { CreateGroupModal } from '../view/sidebar/create-group-modal';
 import { DeleteModal } from '../view/sidebar/delete-modal';
 import { EditModal } from '../view/sidebar/edit-modal/edit-modal';
 import { EmptyModal } from '../view/sidebar/empty-modal';
