+++ src/actions/modals/create-group-modal.tsx
@@ -19,10 +19,10 @@ import { getErrorMessage, Folder } from '@zextras/carbonio-ui-commons';
 import { map } from 'lodash';
 import { useTranslation } from 'react-i18next';
 
-import { MultipleCalendarSelector } from './custom-components/multiple-calendar-selector';
-import { GroupCalendarsList } from './group-calendars-list';
 import { createCalendarGroupRequest } from '../../soap/create-calendar-group-request';
 import { updateCalendarGroupsStore } from '../../store/zustand/calendar-group-store';
+import { MultipleCalendarSelector } from '../../view/sidebar/custom-components/multiple-calendar-selector';
+import { GroupCalendarsList } from '../../view/sidebar/group-calendars-list';
 
 type CreateGroupModalProps = {
 	onClose: () => void;
